backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq==0.9.0
langgraph==0.1.19
langchain-core==0.2.10
langchain-groq==0.1.6
pydantic==2.7.3
pydantic-settings==2.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
httpx==0.27.0
openai==1.35.0
aiofiles==23.2.1
sse-starlette==2.1.0